Multiple changes.
Added a warning pop up message if you set the scan delay to 0. (It can make servers drop some of the port scans)
Added wait cursor when placing the mouse over the progrssbar while the scan is in progress.
Set progressbar value to 0 after cancelling scan.
Disable the delay input box while the scan is in progress.
Deleted duplicated fonts.